pyproject.toml
[tool.poetry]
name = "code-summarizer"
version = "0.1.0"
description = ""
authors = ["AleksandraSorokovikova <alexandraroze2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.40.0"
datasets = "^3.1.0"
ipywidgets = "^8.1.5"
transformers = "^4.46.2"
openai = {extras = ["embeddings"], version = "^1.54.4"}
tiktoken = "^0.8.0"
faiss-cpu = "^1.9.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
rerankers = "^0.6.0"
numpy = "1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
